requirements.txt
pika
python-telegram-bot
fastapi
uvicorn